[Main Group 10] Feat/Navbar

+++ back-end/src/database/models/product.js
@@ -1,7 +1,7 @@
 const Product = (sequelize, DataTypes) => {
   const Product = sequelize.define("product", {
     name: DataTypes.STRING,
-    price: DataTypes.REAL,
+    price: DataTypes.DECIMAL(10,2),
     url_image: DataTypes.STRING,
   }, {
     timestamps: false,
